Use absolute URLs for submodules.
This should make it easier for those wanting to fork this repository
(e.g. on GitHub), by saving them the need to fork the submodules'
repositories as well.

An added bonus: Viewers on GitHub should now be able to follow URL
links in the thirdparty directory, to the submodules' repositories.