Revert "Use git URLs for submodules."
This reverts commit 7a98bac2d3feeafab5680467c0ed3e7c0a18b6d4.

The idea with that commit was to enable pushing changes to the
sub-modules' repos (such as the one recently made to the harfbuzz
repo). This had the side effect of requiring a GitHub account in order
to fully clone the repo, as a user.

Instead, we'll go back to using HTTPS URLs, and I'll set the push
remotes to git URLs locally, for development.