appveyor: Disable building for the VS2019 target, for now
The current Visual Studio release build, tries to build the tesseract
library, as well. The absence of the tesseract sources (and the
leptonica sources, also needed) causes errors in the build, and in spite
of the fact the executables build successfully, the build job fails, and
causes the whole batch to return a failure. Adding the sources doesn't
help, as it causes the build to fail at link-time.

The workaround for now, is to disable the Visual Studio build job, until
either upstream's tesseract integration improves, or another way to test
building with Visual Studio is found.